[OWL-610] Re-implement the commands for NQM
The NQM doesn't have the `control` shell script, so I implement the
`start` and `stop` commands by calling the NQM excutable file directly.
However, the feature which only excutes the newest version is not
support yet. In this commit, I also refined some code and changed the
repo's name to `ops-nqm-agent-updater`.